+++ lib/replace/index.js
@@ -4,7 +4,7 @@ var handlers = {
     AttributeSelector: require('./AttributeSelector'),
     Value: require('./Value'),
     Dimension: require('./Dimension'),
-    Percentage: require('./Dimension'),
+    Percentage: require('./Percentage'),
     Number: require('./Number'),
     String: require('./String'),
     Url: require('./Url'),
